test(dotfiles): root E2E を対象別6ファイルへ分割 [#477]
tests/e2e/mod.rs（1067行）を fzf-picker パターンに揃え、対象別の
6ファイル + ルート mod.rs へ分割する。

- mod.rs はモジュールルートとして doc + mod 宣言 + 横断ヘルパー
  （dotfiles() / write_stub()、pub(crate)）だけを持つ
- cli / apply_copy / list / generate / overlay / claude_settings の
  6ファイルへ各セクションを移設。固有 fixture は各消費ファイルへ、
  current_os は overlay.rs へ
- support.rs は作らない（横断ヘルパーは mod.rs に集約）

純リファクタ。テスト37件は分割前と同数で挙動不変、Cargo.toml は
[[test]] path = "tests/e2e/mod.rs" のまま無変更。
