+++ tests/e2e/cli.rs
@@ -1,0 +1,33 @@
+//! `dotfiles` のトップレベル CLI 契約（`--help` / `--version` / 引数なし）の E2E。
+
+use crate::dotfiles;
+use predicates::prelude::*;
+use rstest::rstest;
+
+#[rstest]
+#[case("--help")]
+#[case("-h")]
+fn help_flag_succeeds(#[case] flag: &str) {
+    dotfiles().arg(flag).assert().success();
+}
+
+#[rstest]
+#[case("--version")]
+#[case("-V")]
+fn version_flag_prints_name_and_version(#[case] flag: &str) {
+    dotfiles()
+        .arg(flag)
+        .assert()
+        .success()
+        .stdout(predicate::str::contains("dotfiles"))
+        .stdout(predicate::str::contains(env!("CARGO_PKG_VERSION")));
+}
+
+#[test]
+fn no_args_prints_version() {
+    dotfiles()
+        .assert()
+        .success()
+        .stdout(predicate::str::contains("dotfiles"))
+        .stdout(predicate::str::contains(env!("CARGO_PKG_VERSION")));
+}
